// jobup-frontend/src/app/generated-sources/openapi/model/chatMessageDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ChatMessageDto { 
    id?: string;
    chatId?: string;
    senderId?: string;
    senderName?: string;
    senderType?: ChatMessageDto.SenderTypeEnum;
    receiverId?: string;
    receiverName?: string;
    receiverType?: ChatMessageDto.ReceiverTypeEnum;
    content?: string;
    messageType?: ChatMessageDto.MessageTypeEnum;
    proposalId?: string;
    dealId?: string;
    createdAt?: string;
}
export namespace ChatMessageDto {
    export type SenderTypeEnum = 'CLIENT' | 'WORKER';
    export const SenderTypeEnum = {
        Client: 'CLIENT' as SenderTypeEnum,
        Worker: 'WORKER' as SenderTypeEnum
    };
    export type ReceiverTypeEnum = 'CLIENT' | 'WORKER';
    export const ReceiverTypeEnum = {
        Client: 'CLIENT' as ReceiverTypeEnum,
        Worker: 'WORKER' as ReceiverTypeEnum
    };
    export type MessageTypeEnum = 'TEXT' | 'PROPOSAL' | 'PROPOSAL_RESPONSE' | 'JOB_CONFIRMED' | 'JOB_COMPLETED' | 'RATING';
    export const MessageTypeEnum = {
        Text: 'TEXT' as MessageTypeEnum,
        Proposal: 'PROPOSAL' as MessageTypeEnum,
        ProposalResponse: 'PROPOSAL_RESPONSE' as MessageTypeEnum,
        JobConfirmed: 'JOB_CONFIRMED' as MessageTypeEnum,
        JobCompleted: 'JOB_COMPLETED' as MessageTypeEnum,
        Rating: 'RATING' as MessageTypeEnum
    };
}